fix distillation
