+++ LuckyBox.ino
@@ -16,6 +16,7 @@
 #include "mashing_mode.h"
 #include "pid_config.h"
 #include "brewing_mode.h"
+#include "misc.h"
 
 void loop() {
   HTTP.handleClient();
@@ -32,27 +33,8 @@ void loop() {
   if (processMode.allow < 3 || processMode.allow > 5) {
 	  adcLoop();
 	  heaterLoop();
-	  // прием мощности по UART
-	  if (Serial.available() > 0) {
-		  uint8_t uartRead = Serial.read();
-		  if (processMode.allow == 1) {
-			  if (processMode.step < 2) power.inPowerHigh = uartRead;
-			  else if (processMode.step < 4) power.inPowerLow = uartRead;
-		  }
-		  else if (processMode.allow == 2) {
-			  if (processMode.step < 2) power.inPowerHigh = uartRead;
-			  else if (processMode.step < 7) power.inPowerLow = uartRead;
-		  }
-	  }
-	  // Выключение повышенного напряжения на клапана
-	  if (CH_all == true) {
-		  CH_all = false;
-
-		  if (timeSetHighVoltage < millis()) {
-			  if (pwmOut[8].invert == false) pwm.setPWM(PWM_CH9, 4096, 0);
-			  else pwm.setPWM(PWM_CH9, 0, 4096);
-		  }
-	  }
+	  serialLoop();
+	  stepApLoop();
   }
 
   sensorLoop();
