+++ misc.h
@@ -1,6 +1,6 @@
 //   Проект центра открытого проектирования у Счастливчика https://LuckyCenter.ru
-#ifndef _PID_CONFIG_h
-#define _PID_CONFIG_h
+#ifndef _MISC_h
+#define _MISC_h
 
 #if defined(ARDUINO) && ARDUINO >= 100
 	#include "arduino.h"
@@ -8,16 +8,13 @@
 	#include "WProgram.h"
 #endif
 
-
 #include "setting.h"
-#include "user_config.h"
-#include "mashing_mode.h"
-#include <time.h>
-#include "misc.h"
 
-extern void initPID();
-extern void loadEepromPid();
-extern void handlePidSet();
-extern void pidSetLoop();
+extern float EEPROM_float_read(int addr);
+extern void EEPROM_float_write(int addr, float val);
+extern void stop_Err();
+extern void check_Err();
+extern void serialLoop();
+extern void stepApLoop();
 
 #endif
