+++ reflux_mode.cpp
@@ -1250,12 +1250,6 @@ void refluxLoop() {
 		}
 	}
 
-	// Выключение повышенного напряжения на клапана
-	/*if (timeSetHighVoltage < millis()) {
-		if (pwmOut[8].invert == false) pwm.setPWM(PWM_CH9, 4096, 0);
-		else pwm.setPWM(PWM_CH9, 0, 4096);
-	}*/
-
 	// Уходим на выбранный алгоритм
 	switch (processMode.number) {
 		case 0: rfluxLoopMode_1(); break; // ручной режим, только сигнализация
